[web] Refactor: detach `AssetRatioSection`
